small changes to client

+++ bank.cpp
@@ -154,7 +154,7 @@ int bank(int cd, const Keys &keys) {
 
 	// put server's ptxt in ctxt
 
-	switch ((int)ptxt[0]) {
+	switch (ptxt[0]) {
 	case BankMsg::DEPOSIT:
 		ctxt[0] = 0;
 	break;
